refactor(governance): 抽 WhitelistObservationMixin 守 R9 行门禁
Task 14 验收:全量 community 套件 test_no_new_oversized_files 红 —
record_process_service.py 因加白名单三路分发超 1000 行(1122)。

把 _handle_whitelist_hit + _refresh_observed_ticket(路2) + _create_observed_ticket
(路3)抽出 record_process_whitelist.py 的 WhitelistObservationMixin,主类继承
(对齐 task_record_query 的 mixin 拆分范式)。import 置于 RecordProcessResult
dataclass 之后避免循环 import(部分加载时该 dataclass 已就绪)。

修复 mixin 抽出时引入的两处 MutableSnapshot 字段名笔误:
estimated_token_saving → estimated_saving_tokens(路2 refresh_snapshot +
路3 MutableSnapshot 构造)。

record_process_service 886 行(达标),mixin 独立文件。
governance+contract+endpoint+architecture 1341 passed 零失败。
